+++ pcsx2/Vif0_Dma.cpp
@@ -170,7 +170,7 @@ __fi void vif0Interrupt()
 
 	g_vif0Cycles = 0;
 
-	vif0Regs.stat.FQC = std::min(vif0ch.qwc, (u16)8);
+	vif0Regs.stat.FQC = std::min(vif0ch.qwc, (u32)8);
 
 	if (!(vif0ch.chcr.STR)) Console.WriteLn("vif0 running when CHCR == %x", vif0ch.chcr._u32);
 
@@ -199,7 +199,7 @@ __fi void vif0Interrupt()
 
 			// One game doesn't like vif stalling at end, can't remember what. Spiderman isn't keen on it tho
 			//vif0ch.chcr.STR = false;
-			vif0Regs.stat.FQC = std::min((u16)0x8, vif0ch.qwc);
+			vif0Regs.stat.FQC = std::min((u32)0x8, vif0ch.qwc);
 			if (vif0ch.qwc > 0 || !vif0.done)
 			{
 				vif0Regs.stat.VPS = VPS_DECODING; //If there's more data you need to say it's decoding the next VIF CMD (Onimusha - Blade Warriors)
@@ -224,7 +224,7 @@ __fi void vif0Interrupt()
 	if (vif0.inprogress & 0x1)
 	{
 		_VIF0chain();
-		vif0Regs.stat.FQC = std::min(vif0ch.qwc, (u16)8);
+		vif0Regs.stat.FQC = std::min(vif0ch.qwc, (u32)8);
 		CPU_INT(DMAC_VIF0, g_vif0Cycles);
 		return;
 	}
@@ -239,7 +239,7 @@ __fi void vif0Interrupt()
 		}
 
 		if ((vif0.inprogress & 0x1) == 0) vif0SetupTransfer();
-		vif0Regs.stat.FQC = std::min(vif0ch.qwc, (u16)8);
+		vif0Regs.stat.FQC = std::min(vif0ch.qwc, (u32)8);
 		CPU_INT(DMAC_VIF0, g_vif0Cycles);
 		return;
 	}
@@ -256,7 +256,7 @@ __fi void vif0Interrupt()
 #endif
 
 	vif0ch.chcr.STR = false;
-	vif0Regs.stat.FQC = std::min((u16)0x8, vif0ch.qwc);
+	vif0Regs.stat.FQC = std::min((u32)0x8, vif0ch.qwc);
 	vif0.vifstalled.enabled = false;
 	vif0.irqoffset.enabled = false;
 	if(vif0.queued_program) vifExecQueue(0);
@@ -307,7 +307,7 @@ void dmaVIF0()
 		vif0.inprogress &= ~0x1;
 	}
 
-	vif0Regs.stat.FQC = std::min((u16)0x8, vif0ch.qwc);
+	vif0Regs.stat.FQC = std::min((u32)0x8, vif0ch.qwc);
 
 	//Using a delay as Beyond Good and Evil does the DMA twice with 2 different TADR's (no checks in the middle, all one block of code),
 	//the first bit it sends isnt required for it to work.
